+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Change Log
 - Renew the snippets feature to enable naming it, changing the order, and performing from the menu.
 - Add the Insert Snippet submenu to the Text menu.
 - Enable the character inspector inspecting more information about each Unicode scalar consists a character.
+- Improve the performance of text editing while the URL detection is enabled.
 - [non-AppStore ver.] Update Sparkle from 2.3.1 to 2.3.2.
 
 
@@ -17,6 +18,11 @@ Change Log
 
 - Reconstruct the Snippets (ex. File Drop) and Key Bindings panes in the Settings window.
 - Update the help contents.
+
+
+### Fixes
+
+- Fix an issue that all URL links removed when the document was updated.